In RelatednessAgg, don't internally round fg/bg_pop
It is appropriate to round relatedness, and to round pop values as
returned to external clients. But internally, it is meaningful to maintain
unrounded pop values.
Esp. for high-cardinality field and bgSet, setting lower min_pop is
indeed meaningful, and rounding arbitrarily restricts this. At a
minimum, there is currently no warning if the user specifies a min_pop
low enough that it will effectively screen out everything.