SOLR-13608: Sync on TrackingBackupRepo NL changes
BackupRepositoryFactory has a special case in its 'newInstance' method.
For 'TrackingBackupRepository' instances, BRF will modify a NamedList
representing the TBF config - adding references to the
SolrResourceLoader and to BRF itself.

In isolation this is innocuous, but becomes a problem when
BRF.newInstance is called concurrently from different threads.  Threads
can race to set these NamedList keys, corrupting the NamedList in the
process.  (i.e. NamedList values ended up in 'key' indices, and vice
versa)

This commit adds synchronization to this NL initialization, to ensure
that the NL values are only set by a single thread.