SOLR-15474: Support Pluggable Circuit Breakers
This commit introduces the capability of adding new circuit breakers on the fly by specifying the
circuit breaker and its parameters in solrconfig.xml.

Previously, it was possible to plug in new circuit breakers by plugging in a custom CircuitBreakerManager.
This commit makes it more elastic by allowing circuit breaker level pluggability.

CircuitBreakerConfig remains to support tests but is no longer used in the core code path.