SOLR-15696: Fix ShardBackupId parsing for backups (#351)
Incremental backups store metadata files whose names are constructed in
part on the name of individual shards.  These filenames are then parsed
later on to extract the shard name and "backup ID".

The parsing in question was overly-brittle though, not accounting for
the variations that shard-splitting etc. produce in shard names.  This
caused incremental backups to fail for any collection that had
previously performed a shard-split.

This commit updates the parsing logic to handle these cases more
flexibly, allowing post-shardsplit backups to now succeed.