SOLR-14834: Update all public-visible links from wiki.apache.org to RefGuide or cwiki (#307)
All references to the old wiki have been updated to point to either the appropriate RefGuide or Cwiki pages.   This doesn't attempt to solve the versioning issues of the Ref Guide.   I've removed a lot of pointers to the 6.6 version of the Ref Guide as well, in favour of the latest Ref Guide, which should hopefully help our SEO on Google point users to the latest version of the docs.