Replace ApiRegistrar method with ApiSupport..getApis
The ApiSupport.getApis method is redundant with the methods used to
register APIs in ApiRegistrar.  I like the clarity of the ApiRegistrar
approach (there's a single place to go to lookup all the APIs available
at a given path), but getApis() is needed for custom requesthandler
definitions to register the v2 APIs at the specified custom path, so
we're stuck with that approach for now.