SOLR-15745: Convert corestatus APIs to annotations
Solr has historically supported two frameworks for definings its v2
APIs: a JSON-based 'apispec' framework, and a more modern replacement
that relies on JAX-RS-like annotations to specify endpoints and
parameters.

The community is in the process of migrating to the annotation based
approach, but many apispec APIs remain.

This commit converts GET /v2/cores and GET /v2/cores/coreName over to
the annotation framework.