SOLR-15501: Read GCS creds more permissively
Prior to this commit, GCSBackupRepository required all users to provide
a path to a file containing GCS credentials.  It turns out that this was
overly strict, as GCP allows hosted code to authenticate implicitly with
whatever roles/permissions assigned to the the hosting server, VM, or pod.
Solr was unintentionally blocking this usecase.

This commit makes the `gcsCredentialPath` setting optional to better
support this usecase.  If the credential path is absent, instead of
throwing an error, a warning is now logged to alert users that they
_might_ be missing this value if they're outside GCP.

Closes: #465
Co-authored-by: Jacek Kikiewicz <public@kikiewicz.com>
Co-authored-by: Martin Stocker <martin.knoller@gmail.com>