SOLR-15501: Fix longstanding GCS test issue
A library we use to stub out GCS in tests only works in a limited set of
locales.  We'd been working around this problem by maintaining a list of
the 'incompatible' locales and skipping the tests if any of the known
bad ones were in use.  This worked much of the time but often failed
because the list of incompatible locales was incomplete and Jenkins jobs
were constantly finding fails with different seeds.

This commit replaces this approach with a different one - have the test
catch the locale-specific error and skip the test at that point.  This
should clean up the sporadic failures that pop up now and then in Jenkins
builds.