SOLR-15745: Create mostly missing core-level req-status API
Prior to this commit the API was mapped in 'CoreApiMapping', but never
existed in the apispec file (and as a result was unusable).

Since the API never fully existed, this commit was forced to take some
guesses at what some of the API details should have been.  SPecifically
I chose here to put the API under
`/cores/someCoreName/command-status/someRequestId`.  I'm unsure whether
this is the best path or not.