SOLR-15964: Transient cores: don't evict open ones (#580)
Transient cores: be careful to not evict a core that is being used, either because of query/indexing, or because the core is undergoing a load/unload/reload operation at that moment.  If that were to happen, it could lead to errors with this core and it being stuck until the node is restarted.

Also:
* Removed CoreContainer methods not needed to be exposed by having the transient cache access the SolrCores from CoreContainer