add a hook for initializing caches by passing a reference to the first registered searcher
This complements (and is analogous to) `SolrCache.warm(...)`; together, the
new `bootstrap(...)` method and the existing `warm(...)` method allow for
consistency across the lifecycle of SolrCache instances.

There is an old comment about possibly passing SolrIndexSearcher in cache
`init(...)`, and having the cache retain a reference to the searcher. We
prefer the approach taken here -- for now -- for backward compatibility,
and so that any change in how SolrIndexSearcher references are passed to
SolrCache instances may be considered as a discrete change.