SOLR-15064: nested docs:  _route_ shouldn't be a fallback for a blank _root_ (#629)
Doing atomic updates for nested documents should not use/assume that the _route_ parameter is a substitute for _root_ if that's not present.  Thus insist the user specify the _root_ field so as to leave no doubt that the atomic update is a child and not itself a root document.

* Rename getChildIdStr -> getSelfOrNestedDocIdStr