SOLR-15870: Fix TestCollectionAPI HTTP error
The error occured because the test was using a single client
to issue Admin commands before and after a server was restarted.
The HTTP Client caches connections, and didn't evict the expired
connection to the server after it restarted. SolrJ does not attempt to
retry Admin calls, therefore the test failed whenever the stale
connection was used.