format via spotlessApply
This is mostly pretty clean; but for SortByFunctionTest in
particular, the diff against main is really opaque. I'm
reasonably certain that what's there now is equivalent
to what's on main. Some of the test spec formatting is
really mangled by spotlessApply (in ways that obscure,
rather than clarify). Is there an annotation or something
to tell spotless to be more tolerant of certain instances?
For now we'll just leave this as spotless wants it.