SOLR-16205: Simplify v2 API span-population logic (#873)
Prior to this commit, V2HttpCall populated the "operationName" of
tracing spans differently depending on whether the API was implemented
as an AnnotatedApi or not.  A special code path for AnnotatedApi's had
the operationName come primarily from the path value specified on the
@Endpoint annotation.

This works well for AnnotatedApis registered at the CoreContainer level,
but actually produces the wrong path for AnnotatedApis registered on
individual cores.  This commit removes this optimized code path, so that
per-core AnnotatedApis produce spans consistent with everything else.