SOLR-15182: Replace emptySpec.json apispec file
Solr's been in the slow process of moving its v2 APIs away from the
existing apispec/mapping framework towards one that relies on more
explicit annotations to specify API properties.

This commit removes emptySpec.json: replacing it with a static constant
in ApiBag.  This is the last apispec file!

This commit does NOT remove any of the framework that allowed parsing
and registering these apispec JSON files.  There should be a good bit of
code that can go, but I wanted to separate its removal into a different
commit for now.