SOLR-15242: Consolidate README.md with solr/README.md (#610)
Consolidate the ./README.md and ./solr/README.md into a new top level README.md optimized for Github.  The /solr/README.adoc is now to help a developer understand what the subdirectories are all about.   Introduced a new README.txt in the ./solr/packaging directory to ship with the distributions.

Co-authored-by: epugh@opensourceconnections.com <>
Co-authored-by: Jan Høydahl <janhoy@users.noreply.github.com>
Co-authored-by: Houston Putman <houston@apache.org>