Allow Jackson in reflection-based serde code
Prior to this commit, Utils.reflectWrite assumed SolrJ's homegrown
"JsonProperty" annotation (created to mimic Jackson).  This allowed
POJOs to be written into maps, but caused problems when trying to use
these same classes with Jersey's out-of-the-box response-writing, which
doesn't know how to interpret our homegrown annotation.

To remedy this, this commit parameterizes Utils.reflectWrite so that it
takes a predicate to filter fields based on an annotation (or any other
criteria).

This allows us to annotate our server-side POJOs with Jackson
annotations, and have them work well with both Jersey serialization (in
v2 code paths), and still reuse the same reflection code to
Map/NamedList-ize them for return by v1 APIs.