Remove ApiRegistrar abstraction
Earlier on during v2 API refactoring, I had created ApiRegistrar as a
way to bundle together the registration of API classes that were
multiplying and cluttering up CoreContainer.load.

In hindsight, this was based on a misunderstanding that there was a
reason to register APIs (directly) in CoreContainer.load in the first
place.  But there's not: both core and container-level APIs can be
registered just fine via the 'getApis()' method present on their
associated requestHandler.  The only APIs that cannot use this approach,
and must be registered explicitly in CoreContainer.load are those v2
APIs that don't have a v1/requestHandler equivalent.

With this in mind, I'm removing my ill-conceived attempt at an
ApiRegistrar class, to simplify the overall picture for API
registration.