SOLR-16346: Generate OpenAPI spec
This commit introduces the 'swagger-gradle-plugin' build-plugin to
Solr's gradle build.  Currently, it's invoked through a new task,
'resolve', that scans the solr-core classpath and generates an OpenAPI
spec based on any detected JAX-RS APIs.

This spec (located in the 'core' module's `build/generated/openapi`
directory) can be used out of the box by any OpenAPI compatible
tooling, such as the code-generator offered by 'swagger' or the web UI
offered by 'Redocly'.  Closer integration with some of this tooling,
particularly as it relates to code-generation, should come in upcoming
commits.