Run "admin" Jersey resources on non-SolrCore hosting nodes
Prior to this commit, our Jersey integration correctly looked up "admin"
APIs whose path contains a core/collection, but only when on nodes that
hosted that core.  Admin requests going to a node that didn't host a
piece of that collection still incorrectly returned a 404, even though
they were registered in the container-level JAX-RS application.

This commit addresses this by adding another "invoke-and-check" case
that makes use of the NotFoundExceptionMapper added by a previous
commit.