ILDPF should resume batches after size violation
Prior to this commit, IgnoreLargeDocumentProcessorFactory only had a
single way to handle documents that exceeded its configurable size
limit.  The first violation would throw a SolrException: in effect,
short-circuiting any other documents in the "batch" being processed.

This approach is very dependent on the ordering of docs within a
batch.  A 100-doc batch with only 1 "size offender" might index 99 docs
or none at all, depending on where the offender is in the list.  This is
great for end users whose clients are built to handle the resulting 400
response, and resubmit their whole batch.  But it's not ideal for every
use case.

This commit introduces an alternate approach for handling these
violations: quietly log out the ID and size of the offending doc but
don't throw any exception that will short-circuit the remainder of the
batch.

The desired error-handling can be chosen using the URP's new config
parameter 'onlyLogErrors'.  When false (the default), the legacy
behavior of short-circuiting the batch and surfacing a 400 error is
used.  Otherwise, the new "just log things out and continue" behavior is
used.