SOLR-16392: Refactor+update v2 DELETEREPLICAPROP API
This commit modifies the existing DeleteReplicaPropertyAPI definition to
be more REST-ful, and use the new JAX-RS framework. Specifically, this
commit changes the API from:

POST /collections/collName
    {"delete-replica-property" : {...}}

to

DELETE /collections/<coll>/shards/<shard>/replicas/<replica>/properties/<prop>

This is a breaking change for users of the v2 API, but one that is
allowed because of v2's "experimental" designation.

In support of this, this commit also refactors DeleteReplicaPropertyAPI
to use the new JAX-RS framework.

Still needed:
  - cleanup unused classes (Payload, etc.)
  - manual testing
  - JUnit tests