SOLR-16392: Support REST-ful path lookup in V2HttpCall
Prior to this commit, V2HttpCall uses a rather crude heuristic to
determine whether a given API is a "container-level" or "core-level"
API.  Namely, any path that a core/collection name can be parsed out of
is a "core" API that should live in the core-level PluginBag.
Conversely, anything else is a container-level API that should live in
the container-level PluginBag.

This happened to work for V2 because we initially chose our API paths
with this heuristic in mind.  But the result is an API that's neither intuitive
nor user-friendly.

This PR takes a sort of "guess and check" approach, where we try
serving ambiguous requests from multiple Jersey apps in series, stopping
at the first one that doesn't spit out a quick NotFoundException.