SOLR-16392:  Refactor and update v2 ADDREPLICAPROP API
This commit modifies the existing AddReplicaPropertyAPI definition to be
more REST-ful, and use the new JAX-RS framework. Specifically, this
commit changes the API from:

POST /collections/collName
    {"add-replica-property" : {...}}

to

PUT /collections/<coll>/shards/<shard>/replicas/<replica>/properties/<prop>
    {"value": <val>}

This is a breaking change for users of the v2 API, but one that is
allowed because of v2's "experimental" designation.

In support of this, this commit also refactors AddReplicaPropertyAPI to
use the new JAX-RS framework.  It also introduces the new class,
"InjectionFactories" to allow Jersey to inject common "context"
values/objects as ctor args for resource classes such as
AddReplicaPropertyAPI.