SOLR-16491: v2 /admin/info/key equivalent, step 6
There's not a ton of logic in PublicKeyHandler that needs refactored
out; the primary bit was a method `createKeyPair` which generates a RSA
key based on some passed in config values.  And the only complicating
issue here is lifecycle: we want to make sure that the key we use in
PublicKeyAPI matches the key used by PublicKeyHandler code, and shares
its lifecycle.

The best way to refactor this code (IMO) was to create a new class that
is primarily responsible for key-generation.  We can then instantiate
this in CoreContainer and make it available to PKAPI via injection.
(This is a small departure from our guess in step 1 that we would want
to make the CloudConfig type injectable, but not a big issue.)

This change bled into a third class, PKIAuthenticationPlugin, which
historically has kept a reference to PublicKeyHandler in order to access
the key pair.  Should we refactor this class to use our new
key-generation utility, or should we let it keep its reference to PKH
for now?

I chose the latter: both out of concerns of scope-creep, and because
PKIAuthPlugin is a public interface that seems more likely than most to
be extended by plugin developers, and there's a slight risk that (e.g.)
changing the ctor signature for that class would cause issues for that
group of users.  Not incredibly likely, and we don't technically promise
backcompat for public classes in solr-core, but it seemed prudent at the
time.