SOLR-16498: Tests to retrieve url from SolrClient (#1149)
Reduce the casting to HttpSolrClient just to get the url out when we can instead access JettySolrRunner object that provides that variable.   Also introduced a getHttpClient() method on SolrJettyTestBase that gets the Apache HttpClient that we use in many tests.  

Started the process of referring to collection1 via the DEFAULT_TEST_COLLECTION_NAME constant.

Co-authored-by: Eric Pugh <epugh@opensourceconnections.com>