DocRouter: strengthen abstraction
DocRouter is an abstraction with 3 implementations; it isn't pluggable.  Years ago at work, I hacked a custom subclass of CompositeIdRouter.  In the process I found a number of spots, especially with splits, that were making assumptions of how CompositeIdRouter parsed IDs.  So here I'm strengthening the separation of concerns so that the split code can do its job, delegating to the DocRouter (specifically a CompositeIdRouter) on how to parse doc IDs.  It also makes CompositeIdRouter more extendable, but I am not truly making the DocRouter choice pluggable because perhaps it's too advanced and would take more work today to do that cleanly & tested.