DocRouter: strengthen abstraction (#1215)
Background: DocRouter is an abstraction with 3 implementations; it isn't pluggable.  There are a number of spots, especially with splits, that were making assumptions of how CompositeIdRouter parsed IDs.  

Strengthen the separation of concerns so that the split code can do its job, delegating to the DocRouter (specifically a CompositeIdRouter) on how to parse doc IDs.  It also makes CompositeIdRouter more extendable, but don't add subclasses or plug-ability.