SOLR-16615: Share Jersey apps across multiple cores
This commit introduces the first pieces needed to share Jersey
applications across cores which share a configset.  This is done by
moving ownership of the Jersey 'ApplicationHandler' from the SolrCore
object itself to a new 'JerseyAppHandlerCache' class.  Reference counts
are used to track how many SolrCores are currently relying on a
particular ApplicationHandler; ApplicationHandlers are removed from the
cache when this reference count hits 0.

In order to support multiple sharing strategies, this commit also
introduces the ConfigsetIdGenerator abstraction.  Each Generator
implementation takes in a 'ConfigSet' and a 'SolrCore' and produces a
String ID that is used to fetch the Jersey application from the cache.

Two implementations are provided: NoSharingConfigsetIdGenerator uses the
SolrCore's UUID as its ID (ensuring that no app-sharing will occur), and
ConfigsetNameIdGenerator uses the name of the configset as its ID (ensuring that
all cores using that configset will share the same Jersey application).

This initial pass has several limitations that need further
investigation:
  - Relying on the configset name alone is insufficient, as configsets
    can be changed "in place" via ZK or filesystem changes.  A more
    nuanced 'ConfigsetIdGenerator' impl will be needed here, perhaps
    that hashes the contents of the configset or something similarly
    content aware?
  - SolrCoreJerseyApp needs new injectors to make sure it can find the
    right SolrCore and metrics objects at request time.