SOLR-16615: Tweak SolrCore injection
The prior mechanism for injecting SolrCore instances into Jersey
resources used a singleton factory, and took advantage of the fact that
each ApplicationHandler was specific to a single core.

This commit replaces the singleton-factory used for injection with a
more standard factory that lifts the SolrCore out of the request
context.