SOLR-16615: Improve "effective configset" resolution
We want to share a Jersey appHandler among all cores that have the same
config, but this requires a deeper check than just looking at the name
of the configset.

The "effective configset" used by a core really combines a number of
different pieces of state: the configset files themselves (primarily
solrconfig.xml), any core or collection properties that are used as
variables in solrconfig.xml, and any config modifications made using the
"overlay" API.

This commit attempts to update ConfigsetIdGenerator to take each of
these chunks of state into account when generating an ID for the
core+configset combination.