SOLR-16573:SolrClientTestRule for EmbeddedSolrServer (#1218)
New SolrClientTestRule abstraction with one implementation: EmbeddedSolrServerTestRule

EmbeddedSolrServerTestBase and AbstractEmbeddedSolrServerTestCase were modified to use this abstraction, and yet were also marked deprecated to encourage direct use of the rule. A solr.xml file is optional. Some affected tests will run more efficiently because (A) they no longer do file I/O to create a temporary solr-home because the rule is able to use existing solr homes and configSets in-place and because (B) initCore() can double-parse a solrconfig.xml and schema.xml.

SolrTestCaseJ4 / SolrTestCase: Moved ObjectReleaseTracker checking to use a Rule so that it's compatible with Rules that use ObjectReleaseTracker.

---------

Co-authored-by: David Smiley <dsmiley@salesforce.com>