Refactor core-admin async logic
Prior to this commit, CoreAdminHandler and CoreAdminAPIBase shared a
decent-sized chunk of very similar logic concerned with the submission
and tracking of async requests.

The logic was non-trivial to deduplicate as v1 and v2 code handle core
admin tasks differently: the v1 code expects a task to essentially
consist of a CoreAdminOp instance wrapped in a 'CallInfo', whereas the
v2 code mostly just expects _something_ to return a SolrJerseyResponse.

This commit attempts to wrestle with that by moving a lot of the common
pieces needed by both v1 and v2 codepaths (MDC setting, map
manipulation, etc.) into CoreAdminAsyncTracker.  With these additional
pieces, CoreAdminAsyncTracker now provides a slightly higher-level
abstraction.  It allows users to submit tasks and mark them as finished,
but doesn't allow the direct manipulation of the underlying status maps.