Change install to trigger restore on all cores
Previously I'd intended 'install' to work by only triggering a restore
on the shard leader, and then relying on Solr's replication to
distribute updates from there.

I think this is still a decent approach, but right now it's not possible
as replication will ultimately fail due to errors that occur at
commit-time due to the collection being in "readOnly" mode.

We should change how readOnly mode works at some point to allow this,
but for now we'll have to do a restore on every replica.