SOLR-16697: Add API to install indices into specific shards (#1458)
API is available at POST /collections/collName/shards/shardName/install

All of the "meat" of this API already existed in some of the utilities used
by the RESTORE API.  This commit largely just adds API bindings to
expose and interface suitable for the single-shard installation use-case.

Callers are expected to put the involved collection into "read only" mode
prior to installing data to shards, and must ensure that the data being
installed is compatible with the collection's schema and config.

(This commit also bundles a few test fixes that were made separately on
'main')