SOLR-16720: Defer PKI header creation to send-time
Prior to this commit, the PKIAuthPlugin decorates outgoing requests at
'queue' time instead of later when they're actually sent.  This can
cause scenarios where, under heavy load, queue time and sending time are
far enough apart that much or all of the TTL window for the outgoing
request is expired by the time its even sent.

This commit switches header-decoration to occur using a different
lifecycle hook offered by Jetty, 'onBegin', so that it's done as close
as possible to the actual time the request is sent.

This won't eliminate the possibility of errant TTL expiry, but should
reduce it substantially under heavy load.