Move 'collection' name into v2 backup path
V2HttpCall has trouble finding the API without this differentiation.

I also think it's more intuitive this way.  The initial thought behind
keeping it out of the path was that once created a backup can be
restored to any collection, so tying it to a single collection in the
path might be a little confusing.  But in hindsight I think the
uniformity we get out of putting the collection in the path makes the
most sense.