SOLR-16720: Defer PKI header creation to send-time (#1495)
Prior to this commit, the PKIAuthPlugin decorates outgoing requests at
'queue' time instead of later when they're actually sent.  This can
cause scenarios where, under heavy load, queue time and sending time are
far enough apart that much or all of the TTL window for the outgoing
request is expired by the time its even sent.

This commit switches header-decoration to occur using a different
lifecycle hook offered by Jetty, 'onBegin', so that it's done as close
as possible to the actual time the request is sent.

This won't eliminate the possibility of errant TTL expiry - a JVM might go
into a long GC right after 'onBegin' for a request but before sending it on
the wire.  But it should reduce errant TTL expiry substantially.