SOLR-16393: Tweak paths of remaining 'POST /collections' APIs
Prior to this commit the v2 APIs for create-collection, create-alias,
and restore-collection were all made available as separate "commands"
under the "POST /collections" API.

This commit changes these APIs to use more differentiated paths:
  - `POST /api/collections` for create-collection (no change)
  - `POST /api/aliases` for create-alias
  - `POST /api/backups/backupName/restore` for restore-collection

This brings the APIs closer (but not fully) into line with their more
REST-ful v2 forms, and allows us to make further API modifications and
convert to JAX-RS individually.

(V2 API calls are dispatched in such a way that the legacy and JAX-RS
framework can never overlap or "share" any of the URL paths that they
serve.  See #1453 for more context.)