SOLR-16395: Avoid Jersey resource inheritance
Prior to this commit GetSchemaFieldAPI inherited from a parent resource
class, GetSchemaAPI.

The "strict validation" linting that Jersey offers (see the config
setting in JerseyApplications.java) complained about this, as it was
detecting that the endpoints offered by GetSchemaAPI were being
shadowed.

This commit removes this inheritance, so that Jersey's validation (when
enabled) can complete successfully on newly created SolrCores.