SOLR-15753: Include entity in v2 request logging
Prior to this commit v2 requests were logged out
PostRequestLoggingFilter, but only the HTTP method, path, and query
parameters were included in the message.  The request body, or entity,
was excluded.

This commit changes this by introducing a special JSON MessageBodyReader
that caches the deserialized request body on Jersey's
"ContainerRequestContext" object, so it can be accessed and logged by
PostRequestLoggingFilter.