SOLR-6994: Integrate post tool into bin/solr and bin/solr.cmd (#1634)
We now support the Post tool on windows by reusing the SolrCLI infrastructure, instead of having a separate post.cmd. This commit does NOT remove the post.jar compilation, or the bin/post shell script.  It does update the Ref Guide to refer to bin/solr post.   In the future we will want to come to a resolution on the post.jar, and maybe rename bin/solr post to something else like bin/solr indexer?   We have very basic bats and unit test now.