Limit files generated by openapi-generator
By default, openapi-generator produces a complete client, including
integration with HTTP libraries, utility classes, documentation, etc.

We want very little of this for our purposes, since we're just looking
to auto-generate SolrRequest implementations for now.  The
'globalProperties' setting allows us to lock down the files that the
generator produces, which really lowers the noise in the generated code.