Add template override directory
We need to make changes to the template for it to look the way we need
it to.  This commit tweaks the generation to rely on a template override
directory, 'src/resources/java-template'.

This will probably need to move eventually - I don't want this included
in the solrj jar after all.  But it's fine for a POC.

This commit doesn't actually make any changes to the api.mustache file
from the default that ships in the generator.  But it sets the stage for
us to make changes in subsequent commits.

In terms of the specific changes, there are two directions we can go in:

  1. Generate request objects that match the current SolrRequest
     contract.  This requires the fewest changes to SolrJ as it exists,
     but requires our generated code to work with NamedList even though
     we have strongly typed classes at our disposal.
  2. Generate request objects that fulfill a SolrRequest contract
     modified in some way to take advantage of the strongly typed
     classes we have available to us.  IMO this would be cleaner
     overall, but would probably require changes to SolrClient, because
     of how the SolrRequest contract is used by SolrClient internally to
     make requests.

I'll start by going the first route and seeing how that turns out.