Ensure 'tidy' cleans up generated code
Modifies the allow/block lists in spotless.gradle to ensure that the
existing 'tidy' task covers the Java code generated by
solr:solrj:openApiGenerate.

Also adds a "finalization" hook to 'openApiGenerate' so that tidy is
run on the generated code straightaway.  (Currently this is done in a
not-quite-ideal way, as we also tidy up non-generated SolrJ code.
That's something I'm hoping to limit/remove in a subsequent commit.)