Restructure packages in solrj-model
The main package is now 'o.a.s.api.', with two subpackages: 'model'
(holding both request and response POJOs), and 'endpoint' (holding the
API interfaces).

This is not just a cosmetic change, but also something that helps
address problems I was having importing certain request/response POJOs
in the API class.  The input to the mustache template tells us which
model classes are in use, but doesn't provide the fully-qualified name
(i.e. the package name is missing).  Putting all models in a single
package makes it easier to create a template that can import these files
from the (unqualified) class name alone.