Move SolrVersion to 'api' module
SolrVersion allows Solr code to parse, compare, and manipulate version
strings.  It also holds a number of constants to indicate the latest and
past versions.  Since it's a conceptual fit for the 'api' submodule, and
because our OpenAPI annotations would benefit from access to these
version constants, this commit moves SolrVersion to the 'api' submodule.

SolrVersion brings with it a dependency on semver4j, so this commit also
adds that dep to 'api' and removes it from 'core'.