Ensure 'api' model classes remain serializable for Javabin
Moving the response POJOs into the 'api' submodule means that they can
no longer declare themselves as MapWriter implementors.  This means that
non-Jackson codecs don't know how to serialize/deserialize these types
for (e.g.) javabin responses.

There were a couple options for solving this:
  1. Move JacksonReflectMapWriter, MapWriter, and related code into
     'api'.
  2. Create a new marker interface that serves essentially the same
     purpose as JacksonReflectMapWriter, but living in 'api'.  Update
     JavaBinCodec etc. to handle this new interface.
  3. Flip flop once again on generating model classes.  The generated
     versions (since they'd only be compiled by SolrJ would have access
     to the necessary interface types).

Not sure this is the best approach, but this commit takes the easiest
option, number 2.