Second pass on serialization headaches
This second pass avoids some of the unfortunate code duplication of the
first pass attempt yesterday.  We're still using a marker interface in
'api' to identify the POJO classes that can be serialized using
reflection, but then we use a new class, DelegateMapWriter, to
serialize them in JavaBinCodec, TextWriter, etc. instead of duplicating
the Utils.reflectWrite call snippet.

Hopefully ends up a little cleaner and more comprehensive.