SOLR-16800: Move towards using the root of Solr as the -solrUrl in the CLI. (#1808)
Now that hostContext can only ever be /solr, and the maturing of the Solr V2 API's that are nested under the /api path, we want Solr commands that take in a -solrUrl to point to the root of your Solr, not a nested path. 

This handles this, as well as provides feedback to users who still use the older -solrUrl http://localhost:8983/solr form that they no longer need to do this.