SOLR-16825: Generate v2 API SolrRequest bindings  (#1793)
Prior to this commit, SolrJ had no real coverage of v2 APIs beyond the
generic V2Request class.  This commit introduces the first request-
specific v2 SolrRequest implementations to SolrJ.

Rather than duplicating details about the API here in SolrJ so that both
the server-side API definition and the client objects need maintained
by hand, this commit generates the v2 SolrRequest classes from the
OpenAPI specification (which in turn is generated from the server-side
API definition).  This ensures that there's only one "source of truth"
that needs maintained for each API.

To achieve this, this commit introduces a new gradle module, 'api', to
hold interfaces (representing Solr's APIs) and POJOs (representing the
API's inputs and outputs).  This module is all that is needed to generate
Solr's OpenAPI spec, and for SolrJ in turn to generate SolrRequest
implementations from that spec.

This commit restricts v2 SolrRequest generation to a few example APIs
to prove the approach.  Additional APIs will be added in subsequent
commits.

---------

Co-authored-by: Houston Putman <houston@apache.org>