SOLR-16733: Fail 'test' tasks using JDK20 on Mac (#1832)
Java 20 has a JDK bug that triggers on Mac systems and crashes Solr on
startup.  This affects deployments running on Mac as well as test runs,
which obviously need to start and stop Solr frequently.

This commit adds some gradle logic to detect this combination (JDK20 w/
Mac) and fail the 'test' task when present.