SOLR-14886 : Suppress stack trace in Query response
Address review comment: read the system property "solr.hideStackTrace"
only once.

The system property is used if the Solr core does not exist or
solr.hideStackTrace is not configured. (in a lot of unit test).
Documentation also mentions that Solr can run without configuring any
core (not sure how useful it is then):
https://solr.apache.org/guide/solr/latest/configuration-guide/core-discovery.html

Pass a Boolean instead of the SolrCore object to methods of
ResponseUtils.

Search the whole code base for places where "trace" is added, and check
for "hideStackTrace" before adding the trace.

Refactor the unit test: Using an Http2SolrClient, the test threw the
exception that was supposed to demonstrate the fix!  Use a
CloseableHttpClient, to query Solr from "outside", without SolrJ, and
get the plain XMl response.