Fix HTML-escaping bug in codegen template
Mustache templating is designed to generate simple HTML pages, and so by
default it escapes input variables as necessary to exist in a HTML
document.

This, combined with the '<' and '>' characters used in Java generics,
causes problems for us in code generation.  To avoid this, this commit
updates our template to use the "triple-brace" syntax in a number of
places, which tells mustache to not do any special escaping of the
templated values.