Add basic auth to PackageTool.
PackageTool actually makes calls to /solr and /api urls, so it actually does the right thing for creating urls.   Supporting this with the SolrCLI.getSolrClient introduced a boolean "barePath" which is set to true for those CLI's that don't just assume a /solr, and actually request a /solr or /api as part of the url.   As we migrate CLI clients, we'll be able to remove this boolean as everyone will assume a barePath.