SOLR-16390: Convert cluster-prop APIs to JAX-RS
Also makes some cosmetic changes to these APIs to bring them a bit more
into line with the REST-ful design we're aiming for going forward.

Still needed for this PR:
  - removal of the old '@Endpoint'-based v2 clusterprop APIs
  - tidy, check, test, etc.

It's worth noting that this PR mostly only aims at parity with existing
functionality:  There are a few other improvements that we _don't_
attempt to solve.  Particularly:
  - it's tough to update anything but top-level property names using the
    single-prop-update API
  - Tough to fetch anything other than top-level properties.
  - explicit 'null' for deletion is still allowed (Should be remove in
    favor of the DELETE API this PR adds)