SOLR-17062: Create JS client for Admin UI use
This uses the 'javascript-closure-angular' generator from the OpenAPI
Generators project.  Not sure if this is the right choice - they also
offer a vanilla 'javascript' option that might be more suitable.

As a JS novice, I'm not in a great position to evaluate the client
itself.  But I can offer it up as a draft PR and see whether others more
familiar with the ecosystem can vet the generated code at all.