Migrate Cross DC code into Solr
- The Cross DC Producer + Cross DC Common -> solr/modules/cross-dc
- The Cross DC Consumer -> solr/cross-dc-manager

We might expand the role of the Cross DC Consumer in the future, so it
felt best to leave it more ambiguous and call it a manager.

The Cross DC Manager relies on the Cross DC Plugin code (for dependency
reasons) so that CrossDC Common code was moved into the Module.