SOLR-16397: Tweak v2 'swapcore' API to be more REST-ful
This commit changes the v2 "swapcore" API to be more in line
with the REST-ful design we're targeting for Solr's v2 APIs.

Following these changes, the v2 API now appears as:
  `POST /api/cores/coreName/swap {...}`

Although not shown above, the 'swap' command specifier
has been removed from the request body.

This commit also converts the API to the new JAX-RS framework.

---------

Co-authored-by: iamsanjay <sanjaydutt.india@yahoo.com>
Co-authored-by: Jason Gerlowski <gerlowskija@apache.org>