WIP: Import and use Solr JS client
NOTE: DOESN'T WORK!  Currently on loading the Admin UI, solr/index.js
triggers a number of import errors.  At a glance, this appears to be
because the generated Javascript file uses a style of import that omits
the file-extension (i.e. '.js').  AFAICT, this extension-less style of
import is common in node.js apps, or when using a 'npm' packaged module,
but doesn't work well when using "unbundled" JS files.

We can change the generation template to include the '.js' suffix in
imports, but I'm not sure if that's the right approach.

Once the import issues are resolved, this commit includes code changes
to switch the Admin UI's collection-deletion over to using the v2 API.
Hard to test or validate this code until the import issues are resolved,
but for now it shows what example usage should looks like.